Added writing files, reading files, listing files, removing files.
I added a very simple writing, reading, listing and deleting file system. 
It writes everything to RAM in an array, so you really shouldn't go much overboard with this.
Commands are simple and are included in the main file. The arduino doesn't really need any other extention or anything, just upload the code and hacker away!